+++ packages/shared-pure/src/index.ts
@@ -1,3 +1,4 @@
 export * from './tools/index.js'
 export * from './types/index.js'
 export * from './utils/index.js'
+export * from './number-format/index.js'
