+++ packages/shared-pure/src/number-format/formatCurrency.ts
@@ -1,6 +1,6 @@
-import type { StringWithAutocomplete } from '@l2beat/shared-pure'
-import { assert } from '@l2beat/shared-pure'
-import { formatNumber } from './formatNumber'
+import type { StringWithAutocomplete } from '../types/StringWithAutocomplete.js'
+import { assert } from '../tools/assert.js'
+import { formatNumber } from './formatNumber.js'
 
 const currencyToSymbol: Record<string, string> = {
   usd: '$',
