+++ packages/shared-pure/src/number-format/formatNumber.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'earl'
 
-import { formatNumber, formatNumberWithCommas } from './formatNumber'
+import { formatNumber, formatNumberWithCommas } from './formatNumber.js'
 
 describe(formatNumber.name, () => {
   describe('default decimals', () => {
